+++ backend/app/controllers/tags_controller.rb
@@ -1,0 +1,9 @@
+class TagsController < ApplicationController
+  def index
+    tags = Tag.order(:name)
+
+    render json: {
+      tags: tags.map { |tag| { id: tag.id, name: tag.name, slug: tag.slug } }
+    }
+  end
+end