+++ backend/app/models/cover_tag.rb
@@ -1,0 +1,4 @@
+class CoverTag < ApplicationRecord
+  belongs_to :cover
+  belongs_to :tag
+end